# swagger.yaml
openapi: 3.0.0
info:
  title: Event Booking API
  version: 1.0.0

paths:

  /events:
    get:
      summary: Get all upcoming events
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

    post:
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, date, total_capacity]
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                total_capacity:
                  type: integer
      responses:
        '201':
          description: Event created
        '400':
          description: Bad request


  /bookings:
    post:
      summary: Book tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, event_id, tickets]
              properties:
                user_id:
                  type: integer
                event_id:
                  type: integer
                tickets:
                  type: integer
      responses:
        '201':
          description: Booking successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  booking_code:
                    type: string
        '400':
          description: Validation error


  /users/{id}/bookings:
    get:
      summary: Get bookings of a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of bookings


  /events/{id}/attendance:
    post:
      summary: Mark attendance using booking code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Attendance recorded
        '400':
          description: Already checked in
        '404':
          description: Invalid code